Update script.js
To remove the following browser console error; The error "ReferenceError: initSizeCtl is not defined" occurred in file "lib/plugins/translate/script.js". If this is in a plugin try updating or disabling the plugin, if this is in a template try updating the template or switching to the "dokuwiki" template.